requirements.txt
django<5.0,>4.2.*
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
django-filter==23.2
django-health-check==3.17.0
drf-spectacular==0.26.2
psycopg2==2.9.6
unittest-xml-reporting==3.2.0
Pillow==9.5.0
factory-boy==3.2.1
coverage==7.2.5
psutil==5.9.5
django-redis-cache==3.0.1
djoser
celery==5.2.7
django-celery-beat==2.5.0
freezegun